// src/app/generated/caster-api/model/createVariableCommand.ts
/**
 * Caster API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { VariableType } from './variableType';


export interface CreateVariableCommand { 
    designId?: string;
    name?: string | null;
    type?: VariableType;
    defaultValue?: string | null;
}